python_backend/requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
google-generativeai>=0.7.2
chromadb==0.4.22
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.6.0
sqlalchemy==2.0.25
numpy<2.0.0
